input command improvements and dev implementation
revised the format for commands to be input. reworked variable name
throughout twitter side code to be more consistent with what's printed
to the console and so they make a bit more sense. global variable now
governs whether the hardware is initialised so twitter side code can be
improved away from ALU hardware.